ai/ai-1-gps/requirements.txt

flask

flask-restx

geopy
requests
joblib
pandas
scikit-learn
python-dotenv
